chore: rename rx modules
