+++ src/snakia/core/rx/combines.py
@@ -6,7 +6,7 @@ from snakia.utils import to_async
 from .async_bindable import AsyncBindable
 from .base_bindable import ValueChanged
 from .bindable import Bindable
-from .concat import concat
+from .concats import concat
 
 A = TypeVar("A")
 B = TypeVar("B")
